# Optimize Active Job argument serialization by ~5x