--- a/activejob/lib/active_job/arguments.rb
+++ b/activejob/lib/active_job/arguments.rb
@@ -82,6 +82,8 @@ module ActiveJob
               argument
             end
           end
+        when Symbol
+          { OBJECT_SERIALIZER_KEY => "ActiveJob::Serializers::SymbolSerializer", "value" => argument.name }
         when GlobalID::Identification
           convert_to_global_id_hash(argument)
         when Array