--- a/activejob/lib/active_job/serializers/action_controller_parameters_serializer.rb
+++ b/activejob/lib/active_job/serializers/action_controller_parameters_serializer.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module ActiveJob
+  module Serializers
+    class ActionControllerParametersSerializer < ObjectSerializer
+      def serialize(argument)
+        Arguments.serialize(argument.to_h.with_indifferent_access)
+      end
+
+      def deserialize(hash)
+        raise NotImplementedError # Serialized as a HashWithIndifferentAccess
+      end
+
+      def serialize?(argument)
+        argument.respond_to?(:permitted?) && argument.respond_to?(:to_h)
+      end
+
+      def klass
+        if defined?(ActionController::Parameters)
+          ActionController::Parameters
+        end
+      end
+    end
+  end
+end