--- a/activejob/lib/active_job/serializers/object_serializer.rb
+++ b/activejob/lib/active_job/serializers/object_serializer.rb
@@ -35,7 +35,8 @@ module ActiveJob
 
       # Serializes an argument to a JSON primitive type.
       def serialize(hash)
-        { Arguments::OBJECT_SERIALIZER_KEY => self.class.name }.merge!(hash)
+        hash[Arguments::OBJECT_SERIALIZER_KEY] = self.class.name
+        hash
       end
 
       # Deserializes an argument from a JSON primitive type.