--- a/activejob/lib/active_job/serializers/object_serializer.rb
+++ b/activejob/lib/active_job/serializers/object_serializer.rb
@@ -17,11 +17,9 @@ module ActiveJob
     #       Money.new(hash["amount"], hash["currency"])
     #     end
     #
-    #     private
-    #
-    #       def klass
-    #         Money
-    #       end
+    #     def klass
+    #       Money
+    #     end
     #   end
     class ObjectSerializer
       include Singleton
@@ -41,15 +39,14 @@ module ActiveJob
       end
 
       # Deserializes an argument from a JSON primitive type.
-      def deserialize(json)
-        raise NotImplementedError
+      def deserialize(hash)
+        raise NotImplementedError, "#{self.class.name} should implement a public #deserialize(hash) method"
       end
 
-      private
-        # The class of the object that will be serialized.
-        def klass # :doc:
-          raise NotImplementedError
-        end
+      # The class of the object that will be serialized.
+      def klass
+        raise NotImplementedError, "#{self.class.name} should implement a public #klass method"
+      end
     end
   end
 end