Fix: https://github.com/rails/rails/pull/55579 (FYI @camertron)

See individual commits for details.

Overall gains (benchmark source is in the first commit)

```
ruby 3.5.0dev (2025-08-27T10:41:07Z master 5257e1298c) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
opt-aj-serialize-index
                        26.282k i/100ms
Calculating -------------------------------------
opt-aj-serialize-index
                        268.003k (± 1.0%) i/s    (3.73 μs/i) -      1.340M in   5.001883s

Comparison:
                  main:    52026.7 i/s
opt-aj-serialize-index:   268003.4 i/s - 5.15x  faster
```

The benchmark is taken from https://github.com/rails/rails/pull/55579. I'm not convinced it's particularly representative of job arguments payloads out there, but it does test a wide range of serializers.

Some stones left unturned:

  - `DateTime` objects are serialized with `TimeSerializer` calling `DateTime#xmlschema` and somehow this is way slower than `Time#xmlschema`. `DateTime` objects are somewhat deprecated, but could be worth checking what's up.
  - The format itself is very inefficient, as discussed with @Earlopain in https://github.com/rails/rails/pull/55544, things like `_aj_symbol_keys` being an array of string is a waste if we assume most hash either have all or none of their keys are symbol. However this requires a schema change, so is a bit out of scope of this PR.
  - The `_aj_serialized` values being fully qualified class name is very wasteful. At least for the builtin ones we should have smaller names. The benchmark payload is `1.35kB` once serialized in JSON, that's a bit ridiculous. It could be shinked by 25% just by removing the namespace part from the serializers, or `~30%` by also using shorter keys.